+++ CHANGELOG.md
@@ -1,7 +1,11 @@
 # Change Log
+
+
+0.0.8 增加blizzard.j 变量的自动补全  前缀  vr 
+0.0.7 增加blizzard.j 常量的自动补全 前缀cn          (constant)
 0.0.6 修改markdown文件,重新定义前缀字符  变量类型前缀:ty  常量前缀:cn 函数前缀:fn 变量前缀:vr
-0.0.4 增加blizzard.j 函数自动补全 前缀fn 修复0.0.4版本中,枚举补全不生效的bug
-0.0.4 增加common.j 枚举的自动补全 前缀nm
-0.0.3 增加common.j 变量的自动补全 前缀vr
+0.0.4 增加blizzard.j 函数自动补全 前缀fn 修复0.0.4版本中,枚举补全不生效的bug    (function)
+0.0.4 增加common.j 常量的自动补全 前缀nm    已经无效
+0.0.3 增加common.j 变量的自动补全 前缀vr    
 0.0.2 增加common.j 函数自动补全 前缀fn
 0.0.1 插件初始化